Finish restructuring query handler.
