+++ components/query_handler/queryHandler.py
@@ -1,47 +1,48 @@
 import json
 from MSG_TYPES import MSG_TYPES
+from helper import getMessageFromCode
 from queryFunctions import *
 
-
-def getMessageFromCode(code):
+def handleQuery (message):
     '''
-        Get the code description from the code.
+        Handle the query.
     '''
-    for key in MSG_TYPES:
-        value = MSG_TYPES[key]
-        if (value == code):
-            return key        
-    return "Unkown message code."
+    print(f"Received message: {getMessageFromCode(message['code'])}")
+    
+    if "code" not in message:
+        response = handleUnknownMessage(message= message)
+    
+    elif (message["code"] == MSG_TYPES["GET_SYSTEMS"]):            
+        response = handleGetSystems(message= message)
+        
+    elif (message["code"] == MSG_TYPES["GET_PROGRAMS"]):            
+        response = handleGetPrograms(message= message)
+    
+    elif (message["code"] == MSG_TYPES["GET_DEPLOYMENTS"]):
+        response = handleGetDeployments(message= message)
+    
+    elif (message["code"] == MSG_TYPES["GET_TRACES"]):
+        response = handleGetTraces(message= message)
+    
+    else:
+        response = handleUnknownCode(message= message)
 
-async def handle_query(websocket):
+    return response
+
+async def receieveMessage(websocket):
     '''
         Handles messages from websocket and echos a response.
     '''
     async for message in websocket:
-        message = json.loads(message)        
-
-        print(f"Received message: {getMessageFromCode(message['code'])}")
-        
-        if "code" not in message:
-            response = handleUnknownMessage(message= message)
-        
-        elif (message["code"] == MSG_TYPES["GET_SYSTEMS"]):            
-            response = handleGetSystems(message= message)
-            
-        elif (message["code"] == MSG_TYPES["GET_PROGRAMS"]):            
-            response = handleGetPrograms(message= message)
-        
-        elif (message["code"] == MSG_TYPES["GET_DEPLOYMENTS"]):
-            response = handleGetDeployments(message= message)
-        
-        elif (message["code"] == MSG_TYPES["GET_TRACES"]):
-            response = handleGetTraces(message= message)
-        
-        else:
-            response = handleUnknownCode(message= message)
-            
-        await websocket.send(json.dumps(response))
-
+        try:
+            message = json.loads(message)   
+            response = handleQuery(message=message)   
+            await websocket.send(json.dumps(response))  
+        except json.JSONDecodeError as e:
+            response = {} 
+            response["response"] = "The query does not contain a valid JSON string."
+            response["error"] = False
+            await websocket.send(json.dumps(response))
